requirements.txt
python-telegram-bot
pyserial